requirements.txt
google-generativeai>=0.8.0
streamlit>=1.30.0
pandas>=2.0.0
python-dotenv>=1.0.0
